requirements.txt
future==0.18.2
numpy==1.16.4
tensorflow==2.1.0
malmo==0.37.0.0
Pillow==7.1.1
